services/ocr-nlp/pyproject.toml
[project]
name = "okemid-ocr-nlp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "python-multipart==0.0.9",
  "pillow==10.4.0",
  "numpy==2.0.2",
  "pdf2image==1.17.0",
  "easyocr==1.7.2",
  "python-dateutil==2.9.0.post0"
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
